change the db structure - remove send_response and feedback table - merge all responses into one table managed by their status
