+++ auto-kol/src/database/schema.sql
@@ -11,11 +11,10 @@ CREATE TABLE IF NOT EXISTS tweets (
     author_username TEXT NOT NULL,
     content TEXT NOT NULL,
     created_at TIMESTAMP NOT NULL,
-    processed_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    engagement_status TEXT CHECK(engagement_status IN ('pending', 'engaged', 'skipped')) NOT NULL
+    processed_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
-CREATE TABLE IF NOT EXISTS pending_responses (
+CREATE TABLE IF NOT EXISTS responses (
     id TEXT PRIMARY KEY,
     tweet_id TEXT NOT NULL,
     content TEXT NOT NULL,
@@ -37,25 +36,6 @@ CREATE TABLE IF NOT EXISTS skipped_tweets (
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (tweet_id) REFERENCES tweets(id)
 );
-
-CREATE TABLE IF NOT EXISTS sent_responses (
-    id TEXT PRIMARY KEY,
-    tweet_id TEXT NOT NULL,
-    response_id TEXT NOT NULL,
-    sent_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    engagement_metrics TEXT, -- JSON string for likes, retweets, etc.
-    FOREIGN KEY (tweet_id) REFERENCES tweets(id),
-    FOREIGN KEY (response_id) REFERENCES pending_responses(id)
-);
-
-CREATE TABLE IF NOT EXISTS feedback (
-    id TEXT PRIMARY KEY,
-    response_id TEXT NOT NULL,
-    feedback_type TEXT CHECK(feedback_type IN ('approve', 'reject', 'improve')) NOT NULL,
-    feedback_content TEXT,
-    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    FOREIGN KEY (response_id) REFERENCES pending_responses(id)
-);
  
 CREATE TABLE IF NOT EXISTS dsn (
     id TEXT PRIMARY KEY,
@@ -66,11 +46,10 @@ CREATE TABLE IF NOT EXISTS dsn (
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (tweet_id) REFERENCES tweets(id),
     FOREIGN KEY (kol_username) REFERENCES kol_accounts(username),
-    FOREIGN KEY (response_id) REFERENCES sent_responses(id)
+    FOREIGN KEY (response_id) REFERENCES responses(id)
 );
 
 CREATE INDEX idx_kor_accounts_username ON kor_accounts(username);
 CREATE INDEX idx_tweets_created_at ON tweets(created_at);
-CREATE INDEX idx_pending_responses_status ON pending_responses(status);
-CREATE INDEX idx_pending_responses_tweet_id ON pending_responses(tweet_id);
+CREATE INDEX idx_responses_status ON responses(status);
 CREATE INDEX idx_skipped_tweets_created_at ON skipped_tweets(created_at);
